Added gitignore, removed unnecessary folder
