--- calculator/calculator.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2013
-VisualStudioVersion = 12.0.31101.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "calculator", "calculator\calculator.csproj", "{1719C8E9-8933-43F4-A2A8-BC666F03FDE7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1719C8E9-8933-43F4-A2A8-BC666F03FDE7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1719C8E9-8933-43F4-A2A8-BC666F03FDE7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1719C8E9-8933-43F4-A2A8-BC666F03FDE7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1719C8E9-8933-43F4-A2A8-BC666F03FDE7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
